+++ solver/src/main/java/org/chocosolver/solver/Model.java
@@ -999,6 +999,22 @@ public class Model implements IModel {
     }
 
     /**
+     * Display for each variable involved in the model, its number of occurrences.
+     */
+    public void displayVariableOccurrences() {
+        Map<String, Integer> l = new HashMap<>();
+        System.out.println("Involved variables: ");
+        for (Variable v : this.getVars()) {
+            l.compute(v.getClass().getSimpleName(), (n, k) -> k == null ? 1 : k + 1);
+        }
+        l.entrySet().stream()
+                .sorted(Map.Entry.comparingByValue())
+                .forEach(e ->
+                        System.out.printf("\t%s (%d)\n", e.getKey(), e.getValue())
+                );
+    }
+
+    /**
      * Display for each propagator involved in the model, its number of occurrences.
      */
     public void displayPropagatorOccurrences() {